+++ portia/execution_agents/default_execution_agent.py
@@ -6,9 +6,8 @@ in completing tasks.
 
 from __future__ import annotations
 
-from typing import TYPE_CHECKING, Any, Literal, cast
+from typing import TYPE_CHECKING, Any, Literal
 
-import instructor
 from langchain_core.messages import SystemMessage
 from langchain_core.prompts import ChatPromptTemplate, HumanMessagePromptTemplate
 from langgraph.graph import END, START, MessagesState, StateGraph
@@ -16,12 +15,13 @@ from langgraph.prebuilt import ToolNode
 from pydantic import BaseModel, ConfigDict, Field, ValidationError
 
 from portia.clarification import Clarification, InputClarification
-from portia.config import EXECUTION_MODEL_KEY, validate_extras_dependencies
+from portia.config import EXECUTION_MODEL_KEY
 from portia.errors import InvalidAgentError, InvalidPlanRunStateError
 from portia.execution_agents.base_execution_agent import BaseExecutionAgent, Output
 from portia.execution_agents.execution_utils import (
     MAX_RETRIES,
     AgentNode,
+    invoke_structured_output,
     next_state_after_tool_call,
     process_output,
     tool_call_or_end,
@@ -34,8 +34,6 @@ from portia.tool import ToolRunContext
 if TYPE_CHECKING:
     from langchain.tools import StructuredTool
     from langchain_core.language_models.chat_models import BaseChatModel
-    from langchain_core.messages import BaseMessage
-    from openai.types.chat import ChatCompletionMessageParam
 
     from portia.config import Config
     from portia.plan import Step
@@ -667,80 +665,3 @@ class DefaultExecutionAgent(BaseExecutionAgent):
             self.tool,
             self.new_clarifications,
         )
-
-
-def invoke_structured_output(
-    model: BaseChatModel,
-    response_model: type[BaseModel],
-    messages: list[BaseMessage],
-) -> dict[Any, Any] | BaseModel:
-    """Invoke a model with structured output.
-
-    This function allows us to dispatch to the structured output method that works with
-    the LLM provider.
-
-    Args:
-        model (BaseChatModel): The LangChain model to invoke.
-        response_model (type[T]): The Pydantic model to use as schema for structured output.
-        messages (list[BaseMessage]): The message input to the model.
-
-    Returns:
-        T: The deserialized Pydantic model from the LLM provider.
-
-    """
-    # We match on class name because not all these types are guaranteed to be installed
-    match model.__class__.__name__:
-        case "ChatOpenAI" | "ChatAnthropic" | "ChatMistralAI" | "AzureChatOpenAI":
-            return model.with_structured_output(
-                response_model,
-                method="function_calling",
-            ).invoke(messages)
-        case "ChatGoogleGenerativeAI":
-            validate_extras_dependencies("google")
-            import google.generativeai as genai
-            if TYPE_CHECKING:
-                from langchain_google_genai import ChatGoogleGenerativeAI
-
-            google_model = cast("ChatGoogleGenerativeAI", model)
-            api_key = (
-                google_model.google_api_key.get_secret_value()
-                if google_model.google_api_key
-                else None
-            )
-            genai.configure(api_key=api_key)  # pyright: ignore[reportPrivateImportUsage]
-            client = instructor.from_gemini(
-                client=genai.GenerativeModel(  # pyright: ignore[reportPrivateImportUsage]
-                    model_name=google_model.model,
-                ),
-                mode=instructor.Mode.GEMINI_JSON,
-            )
-            return client.messages.create(  # pyright: ignore[reportReturnType]
-                messages=_map_message_types_for_instructor(messages),
-                response_model=response_model,
-            )
-        case _:
-            raise ValueError(f"Unsupported model type: {model.__class__.__name__}")
-
-def _map_message_types_for_instructor(
-    messages: list[BaseMessage],
-) -> list[ChatCompletionMessageParam]:
-    """Map the message types to the correct format for the LLM provider.
-
-    Args:
-        messages (list[BaseMessage]): Input Langchain messages
-
-    Returns:
-        list[dict]: The mapped messages.
-
-    """
-    def _map_message(message: BaseMessage) -> ChatCompletionMessageParam:
-        match message.type:
-            case "human":
-                return {"role": "user", "content": message.text()}
-            case "system":
-                return {"role": "system", "content": message.text()}
-            case _:
-                raise ValueError(f"Unsupported message type: {message.type}")
-
-    return [_map_message(message) for message in messages]
-
