+++ portia/execution_agents/execution_utils.py
@@ -6,16 +6,22 @@ This module contains utility functions for managing agent execution flow.
 from __future__ import annotations
 
 from enum import Enum
-from typing import TYPE_CHECKING, Any, Literal
+from typing import TYPE_CHECKING, Any, Literal, cast
 
+import instructor
 from langchain_core.messages import BaseMessage, ToolMessage
 from langgraph.graph import END, MessagesState
 
 from portia.clarification import Clarification
+from portia.config import validate_extras_dependencies
 from portia.errors import InvalidAgentOutputError, ToolFailedError, ToolRetryError
 from portia.execution_agents.base_execution_agent import Output
 
 if TYPE_CHECKING:
+    from langchain_core.language_models.chat_models import BaseChatModel
+    from openai.types.chat import ChatCompletionMessageParam
+    from pydantic import BaseModel
+
     from portia.tool import Tool
 
 
@@ -171,3 +177,80 @@ def process_output(  # noqa: C901
         summaries.append(output.summary or output.serialize_value(output.value))
 
     return Output(value=values, summary=", ".join(summaries))
+
+
+def map_message_types_for_instructor(
+    messages: list[BaseMessage],
+) -> list[ChatCompletionMessageParam]:
+    """Map the message types to the correct format for the LLM provider.
+
+    Args:
+        messages (list[BaseMessage]): Input Langchain messages
+
+    Returns:
+        list[dict]: The mapped messages.
+
+    """
+    def _map_message(message: BaseMessage) -> ChatCompletionMessageParam:
+        match message.type:
+            case "human":
+                return {"role": "user", "content": message.text()}
+            case "system":
+                return {"role": "system", "content": message.text()}
+            case _:
+                raise ValueError(f"Unsupported message type: {message.type}")
+
+    return [_map_message(message) for message in messages]
+
+
+def invoke_structured_output(
+    model: BaseChatModel,
+    response_model: type[BaseModel],
+    messages: list[BaseMessage],
+) -> dict[Any, Any] | BaseModel:
+    """Invoke a model with structured output.
+
+    This function allows us to dispatch to the structured output method that works with
+    the LLM provider.
+
+    Args:
+        model (BaseChatModel): The LangChain model to invoke.
+        response_model (type[T]): The Pydantic model to use as schema for structured output.
+        messages (list[BaseMessage]): The message input to the model.
+
+    Returns:
+        T: The deserialized Pydantic model from the LLM provider.
+
+    """
+    # We match on class name because not all these types are guaranteed to be installed
+    match model.__class__.__name__:
+        case "ChatOpenAI" | "ChatAnthropic" | "ChatMistralAI" | "AzureChatOpenAI":
+            return model.with_structured_output(
+                response_model,
+                method="function_calling",
+            ).invoke(messages)
+        case "ChatGoogleGenerativeAI":
+            validate_extras_dependencies("google")
+            import google.generativeai as genai
+            if TYPE_CHECKING:
+                from langchain_google_genai import ChatGoogleGenerativeAI
+
+            google_model = cast("ChatGoogleGenerativeAI", model)
+            api_key = (
+                google_model.google_api_key.get_secret_value()
+                if google_model.google_api_key
+                else None
+            )
+            genai.configure(api_key=api_key)  # pyright: ignore[reportPrivateImportUsage]
+            client = instructor.from_gemini(
+                client=genai.GenerativeModel(  # pyright: ignore[reportPrivateImportUsage]
+                    model_name=google_model.model,
+                ),
+                mode=instructor.Mode.GEMINI_JSON,
+            )
+            return client.messages.create(  # pyright: ignore[reportReturnType]
+                messages=map_message_types_for_instructor(messages),
+                response_model=response_model,
+            )
+        case _:
+            raise ValueError(f"Unsupported model type: {model.__class__.__name__}")
